requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.11
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
